Moved ResourceList into resources folder
